services/DBpedia_Proxy/requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
blinker==1.4
chardet==3.0.4
click==7.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
Hypercorn==0.10.1
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6
priority==1.3.0
Quart==0.12.0
toml==0.10.1
typing-extensions==3.7.4.2
Werkzeug==1.0.1
wsproto==0.15.0
yarl==1.4.2
pyspotlight==0.7.2